chore: move config merge to a separate file
No code changes.
